+++ cmd/rqlite/history/file.go
@@ -5,19 +5,7 @@ package history
 import (
 	"io"
 	"os"
-	"path/filepath"
 )
-
-const historyFile = ".rqlite_history"
-
-// Path returns the full path to the history file.
-func Path() (string, error) {
-	hdir, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(hdir, historyFile), nil
-}
 
 // Reader returns a reader of the history file.
 func Reader() io.ReadCloser {
@@ -45,13 +33,4 @@ func Writer() io.WriteCloser {
 		return nil
 	}
 	return f
-}
-
-// Delete performs a best-effort removal of the history file.
-func Delete() {
-	p, err := Path()
-	if err != nil {
-		return
-	}
-	os.Remove(p)
 }
